^r Extract cross-command helpers and read-only inspection from authpolicy manage.go, move-only (gofmt, vet, build, package tests pass; declaration set verified identical; final package state byte-matches the verified full-split branch; internal restructure, no behavior change)
Step 3 of 3 for the authpolicy god-file split: shared helpers move
verbatim to manage_helpers.go and the read-only reporting subcommands to
manage_inspect.go, leaving the dispatch surface in manage.go (217 lines).
Also removes the newCmdFlagSet doc comment that step 2 left duplicated in
manage.go (the copy attached to the function in manage_args.go is
canonical).
